backend/requirements/dev.txt
# Include base requirements
-r base.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.10,<5.0
pytest-cov>=5.0,<6.0
pytest-xdist>=3.5,<4.0
factory-boy>=3.3,<4.0
faker>=20.0,<21.0

# Code Quality
black>=24.0,<25.0
flake8>=7.0,<8.0
flake8-bugbear>=24.0,<25.0
isort>=5.13,<6.0
mypy>=1.13,<2.0
django-stubs>=5.1,<5.3
djangorestframework-stubs>=3.15,<3.17

# Debugging
django-debug-toolbar>=4.3,<5.0
ipython>=8.20,<9.0
ipdb>=0.13,<1.0

# Documentation
Sphinx>=7.2,<8.0
sphinx-rtd-theme>=2.0,<3.0

# Pre-commit hooks
pre-commit>=3.6,<4.0

# Load testing
locust>=2.20,<3.0

# Database management
django-extensions>=3.2,<4.0
